Offer option to automatically remove unwanted plugins on launch
This commit allows option to automatically delete unwanted VR plugins.
All user needs to do is to click 'yes'.

Additionally game is checked before launch for being UE title and
warning is issued if it's not.

This requires `--attach=` parameter to be full path to game executable.
If only process name is passed old behavior is used instead.