Add UI explanation and toggle for launching games
Adds 'autostart game' toggle. User can turn it off and `--launch=` will
be ignored.

When `--launch=` option is passed explanation dialog will appear with
option to launch game or not and hide the dialog on subsequent runs.

This allows launchers like Rai Pal to always pass `--launch=` along with
`--attach=`.